create SearchContainer components
